+++ .buildscript/deploy_snapshot.sh
@@ -5,7 +5,7 @@
 # Adapted from https://coderwall.com/p/9b_lfq and
 # http://benlimmer.com/2013/12/26/automatically-publish-javadoc-to-gh-pages-with-travis-ci/
 
-SLUG="arcao/slf4j-timber"
+SLUG="lionzxy/slf4j2-timber"
 JDK="oraclejdk8"
 BRANCH="master"