+++ slf4j2-timber/src/main/kotlin/uk/kulikov/slf4j2/timber/TimberLoggerServiceProvider.kt
@@ -1,4 +1,4 @@
-package uk.kulikov.slf4j.timber
+package uk.kulikov.slf4j2.timber
 
 import org.slf4j.ILoggerFactory
 import org.slf4j.IMarkerFactory